extract makeRun
