+++ lib/makeOprah.js
@@ -2,42 +2,11 @@
 
 const chalk = require('chalk');
 const { log } = require('../lib/utils/logger');
-const { populateConfig, populateSecret } = require('./services');
 const { makeParameterStore } = require('./services/parameter-store/make-parameter-store');
 const { makeSettingsService } = require('./services/settings/make-settings-service');
 const { makeList } = require('./services/make-list');
 const { makeInit } = require('./services/make-init');
-
-const makeRun = ({ stage, interactive, getSettings, parameterStore }) => () => {
-
-  return getSettings()
-  .tap((settings) => {
-    if (!settings.config) { return; }
-
-    const config = settings.config;
-    config.defaults = Object.assign({}, config.defaults, settings.outputs);
-
-    const overrides = config[stage];
-
-    return populateConfig({
-      settings,
-      config,
-      overrides,
-      interactive,
-      parameterStore
-    });
-  })
-  .then(settings => {
-    if (!settings.secret) { return; }
-
-    return populateSecret({
-      settings,
-      secret: settings.secret,
-      interactive,
-      parameterStore
-    });
-  });
-};
+const { makeRun } = require('./services/make-run');
 
 const makeOprah = ({ params }) => {
   const settingsService = makeSettingsService({
