--- lib/services/index.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const { populateConfig } = require('./populate-config');
-const { populateSecret } = require('./populate-secret');
-
-module.exports = {
-  populateConfig,
-  populateSecret
-};
